Refactored eventId from Int to String

+++ app/src/main/java/com/pawlowski/stuboard/presentation/event_details/EventDetailsViewModel.kt
@@ -14,7 +14,7 @@ class EventDetailsViewModel @Inject constructor(
     private val savedStateHandle: SavedStateHandle
 ): ViewModel(), IEventDetailsViewModel {
 
-    private val eventId = savedStateHandle.get<String>("eventId")!!.toInt()
+    private val eventId = savedStateHandle.get<String>("eventId")!!
 
     private val eventResult = getEventDetailsUseCase(eventId)
     private val isRefreshing = eventResult.map { it == null || !it.isFresh }
